Moved... a lot of files.
It should be obvious that I'm still structuring this project,
and that there's still a lot of things that need to be done.

This isn't going to be the final package layout, but i think it looks
somewhat closer to what I want to see in the end product.

More refactoring and reorganizing commits will be on their way soon.